Switch to uint256

+++ contracts/LandStacking.sol
@@ -8,14 +8,14 @@ import "@openzeppelin/contracts/token/ERC20/utils/TokenTimelock.sol";
 contract LandStacking is ERC20, ERC20Permit, ERC20Votes {
     IERC20 public immutable landToken;
 
-    uint public rewardRate = 30e18;
-    uint public immutable startBlock;
-    uint public immutable endBlock;
-    uint public lastUpdateBlock;
-    uint public rewardPerTokenStored;
+    uint256 public rewardRate = 30e18;
+    uint256 public immutable startBlock;
+    uint256 public immutable endBlock;
+    uint256 public lastUpdateBlock;
+    uint256 public rewardPerTokenStored;
 
-    mapping(address => uint) public userRewardPerTokenPaid;
-    mapping(address => uint) public rewards;
+    mapping(address => uint256) public userRewardPerTokenPaid;
+    mapping(address => uint256) public rewards;
     mapping(address => TokenTimelock[]) public timeLocks;
 
     constructor(string memory name_, string memory symbol_, address _landToken)
@@ -27,7 +27,7 @@ contract LandStacking is ERC20, ERC20Permit, ERC20Votes {
         endBlock = block.number + 1e6;
     }
 
-    function rewardPerToken() public view returns (uint) {
+    function rewardPerToken() public view returns (uint256) {
         if (totalSupply() == 0) {
             return rewardPerTokenStored;
         }
@@ -40,7 +40,7 @@ contract LandStacking is ERC20, ERC20Permit, ERC20Votes {
         (((lastBlock - lastUpdateBlock) * rewardRate * 1e18) / totalSupply());
     }
 
-    function earned(address account) public view returns (uint) {
+    function earned(address account) public view returns (uint256) {
         return
         ((balanceOf(account) *
         (rewardPerToken() - userRewardPerTokenPaid[account])) / 1e18) +
@@ -49,7 +49,7 @@ contract LandStacking is ERC20, ERC20Permit, ERC20Votes {
 
     modifier updateReward(address account) {
         rewardPerTokenStored = rewardPerToken();
-        uint lastBlock = block.number;
+        uint256 lastBlock = block.number;
         if (lastBlock > endBlock) {
             lastBlock = endBlock;
         }
@@ -60,12 +60,12 @@ contract LandStacking is ERC20, ERC20Permit, ERC20Votes {
         _;
     }
 
-    function stake(uint _amount) external updateReward(msg.sender) {
+    function stake(uint256 _amount) external updateReward(msg.sender) {
         landToken.transferFrom(msg.sender, address(this), _amount);
         _mint(msg.sender, _amount);
     }
 
-    function withdraw(uint _amount) external updateReward(msg.sender) {
+    function withdraw(uint256 _amount) external updateReward(msg.sender) {
         TokenTimelock timeLock = new TokenTimelock(landToken, msg.sender, block.timestamp + 30 days);
         landToken.transfer(address(timeLock), _amount);
         timeLocks[msg.sender].push(timeLock);
@@ -73,7 +73,7 @@ contract LandStacking is ERC20, ERC20Permit, ERC20Votes {
     }
 
     function getReward() external updateReward(msg.sender) {
-        uint reward = rewards[msg.sender];
+        uint256 reward = rewards[msg.sender];
         rewards[msg.sender] = 0;
         landToken.transfer(msg.sender, reward);
     }
